// workspace/enums.ts
export enum Shares_transactions_FieldIds {
  company_d92 = "68ef35b1ec49f35fffc1bd92",
  to_shareholder_e0c = "68ef35bdec49f35fffc1be0c",
  from_shareholder_e5d = "68ef35ceec49f35fffc1be5d",
  number_of_shares_f45 = "68ef362cec49f35fffc1bf45",
  date_fe1 = "68ef3646ec49f35fffc1bfe1",
  notes_ff1 = "68ef3651ec49f35fffc1bff1",
  from_share_range_eea = "68ef4eaf7d4d64184e6c2eea",
  to_share_range_f81 = "68ef4ebf7d4d64184e6c2f81",
  action_0cc = "68ef4eef7d4d64184e6c30cc",
  is_transaction_valid_627 = "68ef80701982f0531c4e6627",
}

// Phase IDs enum for workflow: Shares transactions
// Use these constants instead of hardcoded IDs in phasesOrder

export enum Shares_transactions_PhaseIds {
  draft_c4b = "68ef3595ec49f35fffc1bc4b",
  pending_627 = "68f5ccb0b7eebea0ebea3627",
  realized_703 = "68f5ccbeb7eebea0ebea3703",
}
// ...
export enum WorkflowIds {
  companies_a51 = "677393544e9c3c0478e35a51",
  tasks_274 = "6773a73ebd84fc36f14d9274",
  time_tracker_ad3 = "6785882fec0d9f8f3f487ad3",
  personnel_815 = "678a1ce0c5ab52aefeaae815",
  materiaalitesitykset_11c = "6799de94f361d9d32544911c",
  ohjausryhmt_2d4 = "67b63ada3188de27bde4e2d4",
  reoccurring_tasks_81f = "68c004aa097b8b294325c81f",
  subtasks_cc4 = "68dd848730628eed73bddcc4",
  kyc_223 = "68e7650aa713093242dbf223",
  contacts_ffb = "68ef3391ec49f35fffc1affb",
  shareholders_13f = "68ef3408ec49f35fffc1b13f",
  shares_transactions_c4c = "68ef3595ec49f35fffc1bc4c",
  marketing_4c3 = "68ef3afe642706a2e322d4c3",
  kyc_signspace_form_6ad = "68f74c6bda248873fd8896ad",
  systems_519 = "690776634ee2003ee256f519",
  systems_in_use_609 = "6907769c4ee2003ee256f609",
  notes_rekna_monthly_1f4 = "69689a250af694ce3738d1f4",
}
// ...
export enum WorkspaceTeams {
  rekna_team_3e3 = "675b24f48a5aeaa8084153e3",
  customer_test_team_a1c = "675b37ee20bedfa2ea6bca1c",
  bots_032 = "698ed2b7d260f553e951b032",
}

// Workspace Groups enum - maps group IDs to their names
// This file is auto-generated by 'hailer-sdk ws-config pull'

export enum WorkspaceGroups {
  customers_767 = "675b372c20bedfa2ea6bc767",
}

// Hailer Members enum - all member types in Hailer ID format
// Format: 'user_{uid}', 'team_{tid}', 'group_{gid}', or 'network_{cid}'
// This file is auto-generated by 'hailer-sdk ws-config pull'

export enum HailerMembers {
  workspace = "network_675b24f38a5aeaa8084153e2",

  // Users
  nico_larsen_74e = "user_5579d000a9bbb6895dbcc74e",
  johan_borgstrm_776 = "user_6540aa56675b6a75feb71776",
  jonathan_wollsten_87f = "user_61263517ca38893eebf9787f",
  philip_ahlgren_98c = "user_675b509d65a81ef3e213598c",
  malin_engblom_637 = "user_61e70e3a6cc1003024700637",
  oliver_nylander_289 = "user_67ee82c3eee6de12e5371289",
  heidi_kuhlberg_662 = "user_68b56662d58a240c32bfe662",
  stefan_atanasov_917 = "user_63bc247911a4ed6dd3519917",
  critix_development_766 = "user_679885a8714b4670dd783766",
  janne_rtknen_6cd = "user_661cc99eb52d2214f8ca66cd",
  susanna_salomaa_d62 = "user_695e081eca61319424bb8d62",
  bot_botinen_03f = "user_698ed2b7d260f553e951b03f",
  darya_petrasheuskaya_dff = "user_5d8b4d357ae3fc0a8fcaedff",

  // Teams
  rekna_team_3e3 = "team_675b24f48a5aeaa8084153e3",
  customer_test_team_a1c = "team_675b37ee20bedfa2ea6bca1c",
  bots_032 = "team_698ed2b7d260f553e951b032",

  // Groups
  customers_767 = "group_675b372c20bedfa2ea6bc767",
}